// src/app/main/shared/dtos/model/catalogo.ts
/**
 * Plataforma Canales Masivos - API Plan Mascotas - Corresponsales
 * Servicios disponibles para el API de Plan Mascotas
 *
 * OpenAPI spec version: 1.0.1
 * Contact: ticanales@suramericana.com.co
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Catalogo { 
    /**
     * Campo al cual desea consultarle el catálogo
     */
    campo: Catalogo.CampoEnum;
}
export namespace Catalogo {
    export type CampoEnum = 'tipoDocumentoAsegurado' | 'genero' | 'formaPago' | 'medioPago' | 'banco' | 'tipoCuenta' | 'tipoDocumentoCuentahabiente' | 'tipoMascota' | 'razaPerro' | 'razaGato' | 'generoMascota' | 'codigoVendedor' | 'ciudadAsegurado' | 'autorizaTratamientoDatos' | 'autorizaEnvioSMS' | 'autorizaEnvioCorreo';
    export const CampoEnum = {
        TipoDocumentoAsegurado: 'tipoDocumentoAsegurado' as CampoEnum,
        Genero: 'genero' as CampoEnum,
        FormaPago: 'formaPago' as CampoEnum,
        MedioPago: 'medioPago' as CampoEnum,
        Banco: 'banco' as CampoEnum,
        TipoCuenta: 'tipoCuenta' as CampoEnum,
        TipoDocumentoCuentahabiente: 'tipoDocumentoCuentahabiente' as CampoEnum,
        TipoMascota: 'tipoMascota' as CampoEnum,
        RazaPerro: 'razaPerro' as CampoEnum,
        RazaGato: 'razaGato' as CampoEnum,
        GeneroMascota: 'generoMascota' as CampoEnum,
        CodigoVendedor: 'codigoVendedor' as CampoEnum,
        CiudadAsegurado: 'ciudadAsegurado' as CampoEnum,
        AutorizaTratamientoDatos: 'autorizaTratamientoDatos' as CampoEnum,
        AutorizaEnvioSMS: 'autorizaEnvioSMS' as CampoEnum,
        AutorizaEnvioCorreo: 'autorizaEnvioCorreo' as CampoEnum
    };
}